# public/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OM Paye API",
        "description": "API pour OM Paye - Système de paiement mobile",
        "version": "1.0.0"
    },
    "paths": {
        "/api/admin/users": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Liste des utilisateurs avec filtres",
                "operationId": "8734f0eb21eaec99b614d902ba09f030",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filtrer par type d'utilisateur",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "admin",
                                "marchand",
                                "utilisateur"
                            ]
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Rechercher par nom, prénom ou téléphone",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de page",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Nombre d'éléments par page",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des utilisateurs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "pagination": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    },
                    "403": {
                        "description": "Accès administrateur requis"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/users/{id}": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Détails d'un utilisateur spécifique",
                "operationId": "fce593e8f474c12b3b475c00ca61f822",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails utilisateur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "type": "object"
                                        },
                                        "compte": {
                                            "type": "object"
                                        },
                                        "transactions": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "statistiques": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Utilisateur non trouvé"
                    },
                    "401": {
                        "description": "Non autorisé"
                    },
                    "403": {
                        "description": "Accès administrateur requis"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Supprimer un utilisateur",
                "operationId": "44eee40ddd4c6892e817da79193968d2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'utilisateur à supprimer",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Utilisateur supprimé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Utilisateur non trouvé"
                    },
                    "401": {
                        "description": "Non autorisé"
                    },
                    "403": {
                        "description": "Accès administrateur requis"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/transactions": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Toutes les transactions avec filtres",
                "operationId": "ffa4d9d0d6b7c1dd523cda0426235a11",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filtrer par type de transaction",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "transfert",
                                "paiement",
                                "depot",
                                "retrait"
                            ]
                        }
                    },
                    {
                        "name": "statut",
                        "in": "query",
                        "description": "Filtrer par statut",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "completed",
                                "failed",
                                "cancelled"
                            ]
                        }
                    },
                    {
                        "name": "date_debut",
                        "in": "query",
                        "description": "Date de début (YYYY-MM-DD)",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_fin",
                        "in": "query",
                        "description": "Date de fin (YYYY-MM-DD)",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de page",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Nombre d'éléments par page",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des transactions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "pagination": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    },
                    "403": {
                        "description": "Accès administrateur requis"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/statistiques": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Statistiques globales du système",
                "operationId": "948f8c365fe88b15fe68a9b596a28eb2",
                "responses": {
                    "200": {
                        "description": "Statistiques globales",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "utilisateurs": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "utilisateurs_normaux": {
                                                    "type": "integer"
                                                },
                                                "marchands": {
                                                    "type": "integer"
                                                },
                                                "admins": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "transactions": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "total_montant": {
                                                    "type": "integer"
                                                },
                                                "transferts": {
                                                    "type": "integer"
                                                },
                                                "paiements": {
                                                    "type": "integer"
                                                },
                                                "depots": {
                                                    "type": "integer"
                                                },
                                                "retraits": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "solde_total": {
                                            "type": "integer"
                                        },
                                        "periode": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    },
                    "403": {
                        "description": "Accès administrateur requis"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token JWT pour l'authentification",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}